Changed stubNewContext from taking a GLint to taking an unsigned long, to allow it to compile and function correctly with 64-bit builds.

+++ opengl_stub/cgl.c
@@ -105,7 +105,7 @@ CGLCreateContext( CGLPixelFormatObj pix, CGLContextObj share, CGLContextObj *ctx
 	if( stub.haveNativeOpenGL )
 		desiredVisual |= ComputeVisBits( pix );
 
-	context = stubNewContext("", desiredVisual, UNDECIDED, (GLint) share);
+	context = stubNewContext("", desiredVisual, UNDECIDED, (unsigned long) share);
 	if (!context)
 		return 0;
